requirements.txt
celery==4.1
colorama>=0.3.9
connexion==1.1.9
cssmin>=0.2.0
elasticsearch>=5.0.0,<6.0.0
eventlet==0.21.0
Flask-Bcrypt==0.6.2
Flask-Cache>=0.13.1
flask-cors>=3.0.2
Flask-Login==0.2.11
Flask-Mail
flask-socketio>=2.9.0
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
flask==0.12.1
flower==0.9.2
gunicorn==19.7
itsdangerous==0.24
Jinja2==2.7.3
jsmin>=2.0.11
markdown
MarkupSafe==0.23
mock==2.0.0
mockredispy==2.9.3
mysqlclient>=1.3.10
newrelic>=2.86.3.70
nose==1.3.7
numpy==1.12.1
psycopg2
pytest-cov==2.4.0
pytest-mock==1.6.0
pytest-xdist==1.15.0
pytest==3.0.7
python-dateutil==2.6.1
redis==2.10.5
requests
rq>=0.7.1
scutils==1.3.0dev2
sqlalchemy==1.1.4
tenacity>=4.1.0
tornado>=4.4.3
Werkzeug==0.10.4
WTForms==2.0.2
pycrypto==2.6.1